requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.63.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
